Fixed history + Detect HDOC

+++ parse_copy.c
@@ -187,7 +187,7 @@ void	create_stack34(t_nlist **stack1, t_nlist **stack2)
 void	create_stack35(t_nlist **stack1, t_nlist **stack2, \
 t_list **stack3, char **s)
 {
-	ft_putendl_fd("HHHHH", 2);
+	//ft_putendl_fd("HHHHH", 2);
 	if (*stack2 != NULL)
 	{
 		*s = join_stack(*stack2);
@@ -424,11 +424,17 @@ char	**gen_args(t_list **lst)
 	nb = 0;
 	while (((char *)(*lst)->next->content)[0] != '>')
 	{
-		array[nb] = ft_strdup((*lst)->content);
+		if (((char *)((*lst)->content))[0] == ' ')
+			array[nb] = ft_substr((*lst)->content, 1, ft_strlen((*lst)->content) - 1);
+		else
+			array[nb] = ft_substr((*lst)->content, 0, ft_strlen((*lst)->content));
 		(*lst) = (*lst)->next;
 		nb++;
 	}
-	array[nb] = ft_strdup((*lst)->content);
+	if (((char *)((*lst)->content))[0] == ' ')
+		array[nb] = ft_substr((*lst)->content, 1, ft_strlen((*lst)->content) - 1);
+	else
+		array[nb] = ft_substr((*lst)->content, 0, ft_strlen((*lst)->content));
 	return (array);
 }
 
@@ -495,6 +501,39 @@ int	only_in(char *str, char c)
 	return (1);
 }
 
+int	only_in2(char *str, char c, char c2)
+{
+	int	i;
+
+	i = -1;
+	while (str[++i])
+	{
+		if (str[i] != c && str[i] != c2)
+			return (0);
+	}
+	return (1);
+}
+
+t_cmd	handle_hdoc(t_list **l, char **env)
+{
+	t_cmd	cmd;
+
+	if (is_in_str('/', (*l)->content) == 0)
+		cmd.path = find_nice_path((*l)->content, env);
+	else
+	{
+		if (((char *)((*l)->content))[0] == ' ')
+			cmd.path = ft_substr((*l)->content, 1, ft_strlen((*l)->content) - 1);
+		else
+			cmd.path = ft_substr((*l)->content, 0, ft_strlen((*l)->content));
+	}
+	*l = (*l)->next->next->next;
+	ft_putendl_fd((*l)->content, 2);
+	cmd.args = gen_args(l);
+	cmd.type = HDOC;
+	return (cmd);
+}
+
 void	ca23(t_list **l, int *i, t_cmd **cmd_array)
 {
 	if (((char *)((*l)->content))[0] == ' ')
@@ -513,27 +552,35 @@ void	ca22(t_list **l, int *i, t_cmd **cmd_array)
 
 void	ca2(t_list **l, int *i, t_cmd **cmd_array, char **env)
 {
-	if (only_in((*l)->content, ' ') == 1 || only_in((*l)->content, '\t') == 1)
-		exit(0);
+	//if (only_in((*l)->content, ' ') == 1 || only_in((*l)->content, '\t') == 1)
+	//	return (ft_putendl_fd("Z", 2), exit(0));
 	ca22(l, i, cmd_array);
 	(*cmd_array)[*i].args = NULL;
 	(*cmd_array)[*i].type = RED_IN;
 	(*l) = (*l)->next;
-	if (is_in_str('/', (*l)->content) == 0)
-		(*cmd_array)[*i + 1].path = find_nice_path((*l)->content, env);
-	else
-		ca23(l, i, cmd_array);
-	if ((*cmd_array)[*i + 1].path == NULL)
+	if (((char *)((*l)->next->content))[0] == '<' && ((char *)((*l)->next->next->content))[0] == '<')
 	{
-		(*cmd_array)[*i + 1].path = ft_strdup((*l)->content);
-		(*cmd_array)[*i + 1].args = NULL;
-		(*cmd_array)[*i + 1].type = NONE;
-		ft_free(gen_args(l));
+			ft_putendl_fd("HDOC Here", 2);
+			(*cmd_array)[*i + 1] = handle_hdoc(l, env);
 	}
 	else
 	{
-		(*cmd_array)[*i + 1].args = gen_args(l);
-		(*cmd_array)[*i + 1].type = CMD;
+		if (is_in_str('/', (*l)->content) == 0)
+			(*cmd_array)[*i + 1].path = find_nice_path((*l)->content, env);
+		else
+			ca23(l, i, cmd_array);
+		if ((*cmd_array)[*i + 1].path == NULL)
+		{
+			(*cmd_array)[*i + 1].path = ft_strdup((*l)->content);
+			(*cmd_array)[*i + 1].args = NULL;
+			(*cmd_array)[*i + 1].type = NONE;
+			ft_free(gen_args(l));
+		}
+		else
+		{
+			(*cmd_array)[*i + 1].args = gen_args(l);
+			(*cmd_array)[*i + 1].type = CMD;
+		}
 	}
 }
 
@@ -583,9 +630,11 @@ t_cmd	*create_args(t_list *l, int i, char **env)
 	int		j;
 
 	cmd_array = ft_calloc(ft_lstsize(l), sizeof(t_cmd));
-	l = l->next;
+	if (((char *)(l->content))[0] == '<')
+		l = l->next;
 	while (l)
 	{
+		ft_putendl_fd(l->content, 2);
 		ca2(&l, &i, &cmd_array, env);
 		l = l->next;
 		while (l && ((char *)(l->content))[0] != '|')
@@ -639,9 +688,11 @@ int	do1cmd(t_cmd *cmds, int flag, char **env)
 		ft_putendl_fd(": command not found", 2);
 		exit(127);
 	}
+	if (cmds[1].type == HDOC)
+		return (ft_putendl_fd("Need to manage HDOC", 2), exit(3), 0);
 	fd[0] = open(cmds[0].path, O_RDONLY);
 	if (fd[0] == -1)
-		exit(1);
+		return (ft_putendl_fd("Error", 2), exit(1), 0);
 	dup2(fd[0], 0);
 	close(fd[0]);
 	fd[1] = open("/dev/stdout", O_WRONLY | O_APPEND | O_CREAT, 0644);
@@ -708,19 +759,23 @@ void	simulpipe(t_cmd **cmd, char **env)
 	tab[0] = 0;
 	while ((*cmd)[tab[0]].type != END)
 	{
-		ft_putendl_fd(":::::", 2);
+		//ft_putendl_fd(":::::", 2);
 		free((*cmd)[tab[0]].path);
 		if ((*cmd)[tab[0]].args)
 			ft_free((*cmd)[tab[0]].args);
 		(tab[0])++;
 	}
+	free(*cmd);
 	exit(WEXITSTATUS(status));
 }
 
 void	apply_flags(t_list *t_lst, int flags[])
 {
 	if (t_lst == NULL)
+	{
+		ft_putendl_fd("....", 2);
 		exit(1);
+	}
 	while (t_lst)
 	{
 		if (((char *)t_lst->content)[0] == '<')
@@ -746,11 +801,16 @@ void	main3(t_list *t_lst[], int flags[], char **env)
 
 	f = 0;
 	t_lst[0] = t_lst[1];
+	ft_putendl_fd("A", 2);
 	apply_flags(t_lst[0], flags);
+	ft_putendl_fd("B", 2);
 	t_lst[0] = t_lst[1];
+	ft_putendl_fd("C", 2);
 	cmds = create_args(t_lst[1], 0, env);
+	ft_putendl_fd("D", 2);
 	ft_lstclear(&t_lst[1], &ft_del);
-	ft_putendl_fd(")))", 2);
+	ft_putendl_fd("E", 2);
+	//ft_putendl_fd(")))", 2);
 	simulpipe(&cmds, env);
 }
 
@@ -777,7 +837,7 @@ void	main2(char *str, char **env, int *flag, int fd)
 		tab = ft_split(str, '|');
 		while (tab[flags[4]] != NULL)
 		{
-			if (is_in_str('<', tab[flags[4]]) == 0)
+			if (tab[flags[4]][0] != '<')
 			{
 				if (flags[4] == 0)
 					tab[flags[4]] = add_str2("< /dev/stdin ", &(tab[flags[4]]));
@@ -813,6 +873,7 @@ void	main2(char *str, char **env, int *flag, int fd)
 		nlstclear(&lst);
 		ft_putendl_fd(str, 2);
 		free(tab);
+		free(str);
 		main3(t_lst, flags, env);
 	}
 	else
@@ -822,55 +883,33 @@ void	main2(char *str, char **env, int *flag, int fd)
 	}
 }
 
-void enable_raw_mode(struct termios *orig_termios) {
-    struct termios raw;
-
-    tcgetattr(STDIN_FILENO, orig_termios);
-    raw = *orig_termios;
-    raw.c_lflag &= ~(ECHO | ICANON);
-    tcsetattr(STDIN_FILENO, TCSAFLUSH, &raw);
-}
-
-void disable_raw_mode(struct termios *orig_termios) {
-    tcsetattr(STDIN_FILENO, TCSAFLUSH, orig_termios);
-}
-
 int	main(int c, char **v, char **env)
 {
 	int		flag;
 	int		fd;
 	char	*str;
-	char	sub[2];
-	struct termios orig_termios;
-	
-	sub[0] = '\0';
-	sub[1] = '\0';
+
 	flag = -1;
-	str = ft_calloc(2, 1);
-	//sub = ft_strdup("");
-	enable_raw_mode(&orig_termios);
-	ft_putstr_fd(">> ", 1);
-	read(STDIN_FILENO, str, 1);
-	//ft_putstr_fd("caracter reached : ", 2);
-	ft_putstr_fd(str, 2);
 	while (1)
 	{
-		while (sub[0] != '\n')
-		{
-			read(STDIN_FILENO, sub, 1);
-			write(STDOUT_FILENO, "\033[2K\r", 5);
-			ft_putstr_fd(">> ", 1);
-			str = add_str(str, sub);
-			//ft_putstr_fd("caracter reached : ", 2);
-			ft_putstr_fd(str, 2);
-		}
-		ft_putendl_fd(str, 2);
-		exit(2);
+		str = readline(">> ");
+		if (str && str[0] != '\0')
+			add_history(str);
 		if (ft_strncmp(str, "exit", 5) == 0)
-			return (disable_raw_mode(&orig_termios), 0);
-		main2(str, env, &flag, fd);
-		unlink("/tmp/tempfile");
-		unlink("/tmp/tempfile2");
+		{
+			free(str);
+			rl_clear_history();
+			return (0);
+		}
+		if (only_in2(str, ' ', '	') == 1)
+			continue ; 
+		else
+		{
+			main2(str, env, &flag, fd);\
+			unlink("/tmp/tempfile");
+			unlink("/tmp/tempfile2");
+			free(str);
+		}
 	}
 	return (1);
 }
